requirements.txt
asn1crypto==0.24.0
autopep8==1.4.4
azure-common==1.1.23
azure-nspkg==3.0.2
azure-storage==0.36.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
pkg-resources==0.0.0
pycodestyle==2.5.0
pycparser==2.19
python-dateutil==2.8.0
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
urllib3==1.25.3
